app/requirements.txt
streamlit
databricks-sql-connector
pandas
plotly
requests
